+++ mimesis/locales.py
@@ -1,9 +1,4 @@
-import contextlib
-from typing import TYPE_CHECKING, Generator
-
-# Enables string-based type annotations
-if TYPE_CHECKING:
-    from mimesis.providers.base import BaseDataProvider
+"""This module provides constants for locale-dependent providers"""
 
 CS = 'cs'
 DA = 'da'
@@ -178,25 +173,3 @@ SUPPORTED_LOCALES = {
 
 LIST_OF_LOCALES = list(SUPPORTED_LOCALES)
 LOCALE_SEPARATOR = '-'
-
-
-@contextlib.contextmanager
-def override(provider: 'BaseDataProvider',
-             locale: str = DEFAULT_LOCALE,
-             ) -> Generator['BaseDataProvider', None, None]:
-    """Context manager which allows overriding current locale.
-
-    :param provider: Locale dependent data provider.
-    :param locale: Locale.
-    :return: Provider with overridden locale.
-    """
-    try:
-        origin_locale = provider.locale
-        provider.override_locale(locale)
-        try:
-            yield provider
-        finally:
-            provider.override_locale(origin_locale)
-    except AttributeError:
-        raise ValueError('«{}» has not locale dependent'.format(
-            provider.__class__.__name__))